Changed PocketSphinx recognition sequence
While working on the Jasper-Melissa merge I noticed how Jasper was using PocketSphinx and tried it out for Melissa. The code is simpler and seems to have better accuracy. My sense is that when the vocabulary is restricted to the command words that it may do quite well.

Moved the decoder initialization to outside and before sphinx_stt so that it is only executed once instead of being in the recognition loop. The first initialization when Melissa starts up takes a few seconds, an inconvenient delay, that might be avoided if the initial user prompt in main.py was done after this initialization. At the moment I do not have a suggestion as to how that would be done. Perhaps the merge will provide a solution.